+++ config-overrides.js
@@ -1,7 +1,7 @@
 const Path = require('path');
 
+// This file configures some "create-react-app" overrides via the "react-app-rewired" package.
 module.exports = {
-  // @ts-ignore
   paths: paths => ({
     ...paths,
     appIndexJs: Path.resolve('./src/frontend/index.tsx'),
