+++ src/cli/buildFrontend.ts
@@ -1,8 +1,11 @@
 import {spawn} from 'child_process';
 
-export function buildFrontend(): Promise<unknown> {
+/**
+ * Builds the frontend of the application using create-react-app.
+ */
+export function buildFrontend(): Promise<void> {
   const child = spawn('yarn react-app-rewired build', {shell: true, stdio: 'inherit'});
   return new Promise((resolve, reject) => {
-    child.on('exit', code => (code ? reject(code) : resolve(code)));
+    child.on('exit', code => (code ? reject(code) : resolve()));
   });
 }
