+++ src/cli/scrape/index.ts
@@ -1,4 +1,4 @@
-import {logger} from '../logger';
+import {logger} from '../../utils/logger';
 import {scrapeDescriptions} from './scrapeDescriptions';
 import {scrapeIndices} from './scrapeIndices';
 import {scrapeKMLs} from './scrapeKMLs';
