+++ src/cli/scrape/scrapeKMLs.ts
@@ -4,9 +4,9 @@ import TJ from '@mapbox/togeojson';
 import xmldom from '@xmldom/xmldom';
 import {cloneDeep, keyBy} from 'lodash';
 import {inspect} from 'util';
-import {logger} from '../logger';
+import {validate} from '../../utils/getValidator';
+import {logger} from '../../utils/logger';
 import cachedFetch from './cachedFetch';
-import {validate} from './getValidator';
 
 /**
  * The maximum number of KMLs to load per batched request.
