+++ src/cli/uploadOutputDir.ts
@@ -1,6 +1,6 @@
 import {S3} from '@aws-sdk/client-s3';
 import {syncS3Dir} from '@scree/aws-utils';
-import {logger} from './logger';
+import {logger} from '../utils/logger';
 import {SyncStackOutput} from './syncStack/getStackTemplate';
 
 export async function uploadOutputDir(s3: S3, outputs: SyncStackOutput) {
