+++ src/utils/logger.ts
@@ -1,7 +1,7 @@
 import chalk from 'chalk';
 import {identity, isString, unzip} from 'lodash';
 import {inspect} from 'util';
-import {OutputStats} from './getOutputStats';
+import {OutputStats} from '../cli/getOutputStats';
 
 /**
  * All messages that get printed to the console should flow through this object.
